move initialization of libs and platforms into new contribution
