+++ arduino-ide-extension/src/browser/arduino-ide-frontend-module.ts
@@ -304,6 +304,7 @@ import { WidgetManager as TheiaWidgetManager } from '@theia/core/lib/browser/wid
 import { StartupTask } from './widgets/sketchbook/startup-task';
 import { IndexesUpdateProgress } from './contributions/indexes-update-progress';
 import { Daemon } from './contributions/daemon';
+import { InitLibsPlatforms } from './contributions/init-libs-platforms';
 
 MonacoThemingService.register({
   id: 'arduino-theme',
@@ -699,6 +700,7 @@ export default new ContainerModule((bind, unbind, isBound, rebind) => {
   Contribution.configure(bind, StartupTask);
   Contribution.configure(bind, IndexesUpdateProgress);
   Contribution.configure(bind, Daemon);
+  Contribution.configure(bind, InitLibsPlatforms);
 
   // Disabled the quick-pick customization from Theia when multiple formatters are available.
   // Use the default VS Code behavior, and pick the first one. In the IDE2, clang-format has `exclusive` selectors.
